Move dates.ts
