+++ services/dailyPageInfo.ts
@@ -1,10 +1,10 @@
-import { nowToday, checkWkday, getJapaneseDate } from "../utils/dates";
 import {
   getDailyMenuURL,
   getTitle,
   getMainText,
   getImageURL,
 } from "../utils/scrapers";
+import { nowToday, checkWkday, getJapaneseDate } from "./dates";
 import { getDailyURL } from "./dailyURL";
 import { sendSlackMessage } from "../utils/webhook";
 import process from "process";
